[REFACTOR] [demo] - Further enforce project conventions.
